+++ apps/sandbox/app/index.tsx
@@ -1,29 +1,34 @@
-import { Link } from "expo-router";
-import React, { ReactNode } from "react";
-import { View, Text } from "react-native";
+import { StyleSheet, Text, View } from "react-native";
 
-export default function Index() {
+export default function Page() {
   return (
-    <View style={{ flex: 1, justifyContent: "center", alignItems: "center" }}>
-      <Text>Home Page!</Text>
-      <StyledLink href="/1">Humpty</StyledLink>
-      <StyledLink href="/1/2">Humpty Dumpty</StyledLink>
-      <StyledLink href="/1/2/3">Humpty Dumpty Sat</StyledLink>
+    <View style={styles.container}>
+      <View style={styles.main}>
+        <Text style={styles.title}>Hello World</Text>
+        <Text style={styles.subtitle}>This is the first page of your app.</Text>
+      </View>
     </View>
   );
 }
 
-const StyledLink = ({
-  href,
-  children,
-}: {
-  href: string;
-  children: ReactNode;
-}) => (
-  <Link
-    href={href}
-    style={{ fontSize: 20, marginVertical: 10, marginHorizontal: 20 }}
-  >
-    {children}
-  </Link>
-);
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    alignItems: "center",
+    padding: 24,
+  },
+  main: {
+    flex: 1,
+    justifyContent: "center",
+    maxWidth: 960,
+    marginHorizontal: "auto",
+  },
+  title: {
+    fontSize: 64,
+    fontWeight: "bold",
+  },
+  subtitle: {
+    fontSize: 36,
+    color: "#38434D",
+  },
+});
